+++ tests/rules/test_transform.py
@@ -201,7 +201,7 @@ def test_transform_model_values_to_editor_values(
     model: AnyExtractedModel | AnyMergedModel | AnyAdditiveModel,
     field_name: str,
     subtractive: AnySubtractiveModel,
-    expected: EditorValue,
+    expected: list[EditorValue],
 ) -> None:
     editor_value = _transform_model_values_to_editor_values(
         model, field_name, subtractive
@@ -516,7 +516,7 @@ def test_transform_models_to_fields() -> None:
 
     assert len(editor_fields) == len(MERGEABLE_FIELDS_BY_CLASS_NAME["MergedPerson"])
     fields_by_name = {f.name: f for f in editor_fields}
-    assert fields_by_name["givenName"].dict() == {
+    assert fields_by_name["givenName"].model_dump() == {
         "is_required": False,
         "value_type": ["str"],
         "name": "givenName",
@@ -583,7 +583,7 @@ def test_transform_models_to_fields() -> None:
             },
         ],
     }
-    assert fields_by_name["memberOf"].dict() == {
+    assert fields_by_name["memberOf"].model_dump() == {
         "is_required": False,
         "value_type": ["MergedOrganizationalUnit"],
         "name": "memberOf",
@@ -734,12 +734,12 @@ def test_transform_fields_to_additive(
                         editor_values=[],
                         name=EditorValue(text="Prevented Primary Source"),
                         identifier=MergedPrimarySourceIdentifier(
-                            "preventedPrimarySourceId"
+                            "preventedPrimarySrcId"
                         ),
                     ),
                 ],
             ),
-            {"familyName": ["preventedPrimarySourceId"]},
+            {"familyName": ["preventedPrimarySrcId"]},
         ),
     ],
 )