+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- change the docker build to multi-stage build
 - change primary source of rules from mex to mex-editor
 - moved pytest rerun config from toml to makefile