+++ CHANGELOG.md
@@ -44,9 +44,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- upgrade reflex to 0.8.15
+- use pydantic v2 BaseModel instead of rx.Base
 - other rerun lib for pytest
 - updated mex-common to 1.16.0
-- upgrade reflex to 0.8.14
 - updated template to https://github.com/robert-koch-institut/mex-template/commit/81bc5b
 
 ### Removed